6th Sem - COMP 342/mini_project_graphics/requirements.txt
glfw==2.5.3
multipledispatch==0.6.0
numpy==1.22.3
Pillow==9.1.1
PyOpenGL==3.1.6
pyrr==0.10.3
six==1.16.0
